move POMs to their respective folders
