+++ tests/e2e/domains/beratungshilfe/formular/BeratungshilfeFormular.ts
@@ -1,4 +1,4 @@
-import { Formular } from "./Formular";
+import { Formular } from "tests/e2e/domains/shared/Formular";
 
 export class BeratungshilfeFormular extends Formular {
   readonly url = "/beratungshilfe/antrag";
