+++ tests/e2e/domains/beratungshilfe/vorabcheck/BeratungshilfeVorabcheck.ts
@@ -1,4 +1,4 @@
-import { Vorabcheck } from "./Vorabcheck";
+import { Vorabcheck } from "tests/e2e/domains/shared/Vorabcheck";
 
 export class BeratungshilfeVorabcheck extends Vorabcheck {
   readonly url = "/beratungshilfe/vorabcheck";
