+++ tests/e2e/domains/fluggastrechte/formular/FluggastrechteFormular.ts
@@ -1,4 +1,4 @@
-import { Formular } from "./Formular";
+import { Formular } from "tests/e2e/domains/shared/Formular";
 
 export class FluggastrechteFormular extends Formular {
   readonly url = "/fluggastrechte/formular";
