+++ tests/e2e/domains/fluggastrechte/formular/startFluggastrechteFormular.ts
@@ -1,5 +1,5 @@
 import { type Page, expect } from "@playwright/test";
-import type { FluggastrechteFormular } from "tests/e2e/pom/FluggastrechteFormular";
+import type { FluggastrechteFormular } from "tests/e2e/domains/fluggastrechte/formular/FluggastrechteFormular";
 import { expectPageToBeAccessible } from "tests/e2e/util/expectPageToBeAccessible";
 import { today, toGermanDateFormat } from "~/util/date";
