+++ tests/e2e/domains/fluggastrechte/vorabcheck/FluggastrechteVorabcheck.ts
@@ -1,4 +1,4 @@
-import { Vorabcheck } from "./Vorabcheck";
+import { Vorabcheck } from "tests/e2e/domains/shared/Vorabcheck";
 
 export class FluggastrechteVorabcheck extends Vorabcheck {
   readonly url = "/fluggastrechte/vorabcheck";
