+++ tests/e2e/domains/fluggastrechte/vorabcheck/vorabcheckFunnelCheck.ts
@@ -1,6 +1,6 @@
 import { type Page, expect } from "@playwright/test";
-import { CookieSettings } from "../../../pom/CookieSettings";
-import type { FluggastrechteVorabcheck } from "../../../pom/FluggastrechteVorabcheck";
+import type { FluggastrechteVorabcheck } from "tests/e2e/domains/fluggastrechte/vorabcheck/FluggastrechteVorabcheck";
+import { CookieSettings } from "../../shared/CookieSettings";
 
 export async function startFluggastrechteVorabcheckFunnelCheck(
   page: Page,
