+++ tests/e2e/domains/geldEinklagen/formular/GeldEinklagenFormular.ts
@@ -1,4 +1,4 @@
-import { Formular } from "./Formular";
+import { Formular } from "tests/e2e/domains/shared/Formular";
 
 export class GeldEinklagenFormular extends Formular {
   readonly url = "/geld-einklagen/formular";
