+++ tests/e2e/domains/geldEinklagen/vorabcheck/GeldEinklagenVorabcheck.ts
@@ -1,4 +1,4 @@
-import { Vorabcheck } from "./Vorabcheck";
+import { Vorabcheck } from "tests/e2e/domains/shared/Vorabcheck";
 
 export class GeldEinklagenVorabcheck extends Vorabcheck {
   readonly url = "/geld-einklagen/vorabcheck";
