+++ tests/e2e/domains/geldEinklagen/vorabcheck/geldeinklagen.vorabcheck.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from "@playwright/test";
-import { CookieSettings } from "tests/e2e/pom/CookieSettings";
-import { GeldEinklagenVorabcheck } from "tests/e2e/pom/GeldEinklagenVorabcheck";
+import { CookieSettings } from "tests/e2e/domains/shared/CookieSettings";
+import { GeldEinklagenVorabcheck } from "tests/e2e/domains/geldEinklagen/vorabcheck/GeldEinklagenVorabcheck";
 import { expectPageToBeAccessible } from "tests/e2e/util/expectPageToBeAccessible";
 
 let geldEinklagen: GeldEinklagenVorabcheck;
