+++ tests/e2e/domains/prozesskostenhilfe/formular/ProzesskostenhilfeFormular.ts
@@ -1,4 +1,4 @@
-import { Formular } from "./Formular";
+import { Formular } from "tests/e2e/domains/shared/Formular";
 
 export class ProzesskostenhilfeFormular extends Formular {
   readonly url = "/prozesskostenhilfe/formular";
