+++ tests/e2e/domains/prozesskostenhilfe/formular/gesetzlicheVertretung.ts
@@ -1,5 +1,5 @@
 import { type Page } from "@playwright/test";
-import type { Formular } from "tests/e2e/pom/Formular";
+import type { Formular } from "tests/e2e/domains/shared/Formular";
 import { expectPageToBeAccessible } from "tests/e2e/util/expectPageToBeAccessible";
 
 export async function startGesetzlicheVertretung(
