+++ tests/e2e/domains/shared/CookieSettings.ts
