+++ tests/e2e/pages/cookie-consent.spec.ts
@@ -2,7 +2,7 @@ import { Buffer } from "buffer";
 import { test, expect } from "@playwright/test";
 import { acceptCookiesFieldName } from "~/components/cookieBanner/CookieBanner";
 import { consentCookieName } from "~/services/analytics/gdprCookie.server";
-import { CookieSettings } from "../pom/CookieSettings";
+import { CookieSettings } from "../domains/shared/CookieSettings";
 import { testPageToBeAccessible } from "../util/testPageToBeAccessible";
 
 const pageUrl = "/cookie-einstellungen";
