+++ tests/e2e/pages/courtfinder.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from "@playwright/test";
-import { CookieSettings } from "../pom/CookieSettings";
-import { CourtFinder } from "../pom/CourtFinder";
+import { CourtFinder } from "./CourtFinder";
+import { CookieSettings } from "../domains/shared/CookieSettings";
 import { expectPageToBeAccessible } from "../util/expectPageToBeAccessible";
 
 let courtfinder: CourtFinder;
